change calculation of svyPRE for multinomial models
This calculates predictions manually for svrepstatmisc objects. The previous version tried to run model with analytical weights using nnet::multinom() to generate an object that worked with the predict() function. However, multinom() couldn't find the weight vector in the function environment, which I tried to solve using global assignment to create a weight vector for the function to call and then removing the weight vector object, which worked using a test function but not the package function.

So, I calculated the predictions manually, which I should have done in the first place. :/